nlu_pipeline/requirements.txt
openai>=1.49.0
numpy>=1.24.0
PyYAML>=6.0.1
# Optional for stronger baseline
# scikit-learn>=1.5.0
